Several changes made.
	• Attempt to parallelize prinseq
	• Several typos/bug fixes
	• Improved SE support in AncientQC

+++ BWAalnMapping.sh
@@ -97,7 +97,7 @@ DeduplicateArray(){ # Deduplicating the sample names
 	#echo "$array"
 	#local sampleNames=$(for name in ${array[@]}; do tmp=$(echo ${name/%%_*}); echo ${tmp/%.f*}; done) # Getting only the sample names
 	local sampleNames=$(for name in ${array[@]}; do
-		tmp="$(echo $name |sed -e 's/_r1.*//I' -e 's/_r2.*//I' -e "s/_merged.*//I" -e "s/\.fa.*//I" -e "s/\.fn.*//I")";
+		tmp="$(echo $name |sed -e 's/_r1.*//I' -e 's/_r2.*//I' -e "s/_merged.*//I" -e "s/\.f.*//I")";
 		echo $tmp;
        	done) # Getting only the sample names
 	#local sampleNames=$(for tmp in ${array[@]}; do tmp=$(echo ${tmp/_r1}); tmp=$(echo ${tmp/_r2}); tmp=$(echo ${tmp/_merged});echo ${tmp/%.f*}; done) # Getting only the sample tmps
@@ -167,7 +167,26 @@ alnMapping(){ # BWA aln Mapping.  Automatically determines if merged or paired.
 		fi
 	fi
 	
-	if [ "$r1" != "NA" ]; then # If I found a Paired set
+	if [[ "$r1" != "NA" && "$r2" == "NA" ]]; then # If I found a merged file
+		bwa aln -o 2 -n 0.01 -l 16500 $ref $r1 -t $ncores > tmp.sai
+		bwa samse $ref tmp.sai $r1 |\
+			samtools view -b -h -F 4 -m $len -q $qual -U tmpSBad.bam |\
+			samtools sort -	> tmpS.bam  
+	
+		samtools fastq tmpS.bam | gzip > ${out}UnmappedReads/${sample}_Single.fastq.gz
+
+		if [ "$multi" != "TRUE" ]; then
+			samtools fastq tmpSBad.bam | gzip > ${out}UnmappedReads/${sample}.fastq.gz
+
+		else
+			# Apr 20 2021 -- Want to separate the poor hits from the reads with multiple hits
+			samtools fastq -f 4 tmpSBad.bam | gzip > tmpUnmapped.fastq.gz #${out}UnmappedReads/${sample}.fastq.gz # All the unmapped Reads
+			samtools view -F 4 tmpSBad.bam | awk '{if($5 > 0){print "@"$1"\n"$10"\n+\n"$11}}' | gzip -c | cat tmpUnmapped.fastq.gz -  > ${out}UnmappedReads/${sample}_r1.fastq.gz # The reads which were poor matches 
+			samtools view -m $len -F 4 tmpSBad.bam | awk '{if($5 == 0){print "@"$1"\n"$10"\n+\n"$11}}' | gzip > ${out}Qual0Reads/${sample}_r1.fastq.gz # The Multimappers
+		fi
+	fi
+
+	if [ "$r1" != "NA" ] && ["$r2" != "NA" ]; then # If I found a Paired set
 	#if [ -v r1 ]; then # If we have paired reads
 		bwa aln -o 2 -n 0.01 -l 16500 $ref $r1 -t $ncores  > tmp1.sai # First index
 		bwa aln -o 2 -n 0.01 -l 16500 $ref $r2 -t $ncores  > tmp2.sai # Second Index
@@ -194,16 +213,20 @@ alnMapping(){ # BWA aln Mapping.  Automatically determines if merged or paired.
 	
 	# Merging the files if needed.  Otherwise, we can ignore and simply mv it
 	#if [ -v merged ] && [ -v r1 ]; then
-	if [ "$merged" != "NA" ] && [ "$r1" != "NA" ]; then
+	if [ "$merged" != "NA" ] && [ "$r1" != "NA" ] && [ "$r2" != "NA" ]; then
 		#echo "MERGED and PAIRED"
 		samtools merge -f ${out}MappedReads/$sample.bam tmpM.bam tmpP.bam
 		samtools merge -f ${out}UnmappedBam/$sample.bam tmpMBad.bam tmpPBad.bam
 		rm tmpP* tmpM*
-	elif [ "$merged" != "NA" ] && [ "$r1" == "NA" ]; then
+	elif [ "$merged" != "NA" ] && [ "$r1" == "NA" ] && [ "$r2" == "NA" ]; then
 	#elif [ -v merged ] && [ -z ${r1+x} ]; then
 		#echo "MERGED"
 		mv tmpM.bam ${out}MappedReads/$sample.bam 
 		mv tmpMBad.bam ${out}UnmappedReads/$sample.bam 
+	elif [ "$merged" == "NA" ] && [ "$r1" != "NA" ] && [ "$r2" == "NA" ]; then
+	#elif [ -v merged ] && [ -z ${r1+x} ]; then
+		mv tmpS.bam  ${out}MappedReads/$sample.bam 
+		mv tmpSBad.bam ${out}UnmappedReads/$sample.bam 
 	else
 		#echo "PAIRED"
 		mv tmpP.bam ${out}MappedReads/$sample.bam
